+++ product_code/user_service/api/src/main/scala/de/upb/cs/uc4/user/api/UserService.scala
@@ -6,13 +6,12 @@ import com.lightbend.lagom.scaladsl.api.broker.Topic
 import com.lightbend.lagom.scaladsl.api.deser.MessageSerializer
 import com.lightbend.lagom.scaladsl.api.transport.Method
 import com.lightbend.lagom.scaladsl.api.{ Descriptor, Service, ServiceAcl, ServiceCall }
-import de.upb.cs.uc4.authentication.model.JsonUsername
 import de.upb.cs.uc4.shared.client.UC4Service
 import de.upb.cs.uc4.shared.client.kafka.EncryptionContainer
 import de.upb.cs.uc4.shared.client.message_serialization.CustomMessageSerializer
 import de.upb.cs.uc4.user.model.post.PostMessageUser
 import de.upb.cs.uc4.user.model.user.{ Admin, Lecturer, Student, User }
-import de.upb.cs.uc4.user.model.{ GetAllUsersResponse, JsonRole, MatriculationUpdate, Usernames }
+import de.upb.cs.uc4.user.model.{ GetAllUsersResponse, JsonRole, MatriculationUpdate }
 
 /** The UserService interface.
   *
@@ -25,6 +24,8 @@ trait UserService extends UC4Service {
   /** The name of the service */
   override val name = "user"
   override val autoAcl: Boolean = false
+
+  val enrollmentIdSecretByteLength: Int = 15
 
   // USER
   /** Get a specific user with the username */