+++ product_code/user_service/api/src/main/scala/de/upb/cs/uc4/user/model/user/Student.scala
@@ -10,6 +10,7 @@ import scala.concurrent.{ ExecutionContext, Future }
 
 case class Student(
     username: String,
+    enrollmentIdSecret: String,
     role: Role,
     address: Address,
     firstName: String,
@@ -23,6 +24,7 @@ case class Student(
 
   def copyUser(
       username: String = this.username,
+      enrollmentIdSecret: String = this.enrollmentIdSecret,
       role: Role = this.role,
       address: Address = this.address,
       firstName: String = this.firstName,
@@ -31,7 +33,7 @@ case class Student(
       phoneNumber: String = this.phoneNumber,
       birthDate: String = this.birthDate
   ): Student =
-    copy(username, role, address, firstName, lastName, email, phoneNumber, birthDate)
+    copy(username, enrollmentIdSecret, role, address, firstName, lastName, email, phoneNumber, birthDate)
 
   override def trim: Student =
     super.trim.asInstanceOf[Student].copy(