+++ product_code/user_service/CHANGELOG.md
@@ -1,5 +1,8 @@
 # [WIP - v0.11.1](https://github.com/upb-uc4/University-Credits-4.0/compare/user-v0.11.0...user-v0.11.1) (2020-XX-XX)
 ## Feature
+ - Added governmentId to PostMessageUser
+ - Added enrollmentIdSecret to user object
+ - Changed default users accordingly, default Users govId is "governmentIdStudent", enrollmentIdSecret is base64 encoding of "studentstudent", "lecturerlecturer" or "adminadmin"
 ## Refactor
  - Changed email addresses, first names and last names for the default users
 ## Bugfix